+++ NEWS.md
@@ -8,6 +8,8 @@ Versioning](https://semver.org/spec).
 
 ### Added
 
+- Arrays of numbers can be efficiently multiplied or divided by neutral numbers.
+
 - Ranges can be constructed with neutral numbers specified as the start, step, and/or stop
   parameters of the range. `𝟙:stop` is identical to `Base.OneTo(stop)` if `stop` is a
   non-neutral integer or is `𝟙` and is identical to `Base.OneTo(Int(stop))` otherwise.